Test case insensitive class names
Class names really should be case sensitive, but I can see a use-case where a dispatcher wants to new up a class based on a URL and that really won’t be strict. In that sense this unit test is just a placeholder for that functionality so it doesn’t get broken in the future.